+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Fixed Aspect icons disapearing during runs
+- Replaced 'Siege Shredder' with 'Mirror Shield' Hammer for Axe Aspect of Young Mel
+
 
 ## [0.2.1] - 2026-04-27